utxo explorer api

+++ apps/explorer/priv/repo/migrations/20241219221925_create_utxoaddressbalance.exs
@@ -4,9 +4,9 @@ defmodule Explorer.Repo.Migrations.CreateUTXOAddressBalance do
   def change do
     create table(:utxoaddressbalance, primary_key: false) do
       add(:address, :string, null: true)
-      add(:available, :bigint, null: false)
-      add(:unavailable, :bigint, null: false)
-      add(:spent, :bigint, null: false)
+      add(:available, :numeric, precision: 100, null: false)
+      add(:unavailable, :numeric, precision: 100, null: false)
+      add(:spent, :numeric, precision: 100, null: false)
       timestamps(null: false, type: :utc_datetime_usec)
 
     end
